Fix Firefox 18 save_url() / nsIWebBrowserPersist.saveURI compatibility.
Since Firefox 18, we need to provide a proper privacyContext
This is cobbled together from half-documented parts in various places
of the Mozilla documentation. The changes file does not list the
necessary steps :-(

https://developer.mozilla.org/en-US/docs/Supporting_per-window_private_browsing
The documentation is even wrong. It recommends to import("chrome://gre/modules/PrivateBrowsingUtils.jsm")
but the correct URL is "resource://gre/modules/PrivateBrowsingUtils.jsm".
Also, the method is not named "getPrivacyContextFromWindow" but "privacyContextFromWindow".

Perls strategy of providing backwards compatibility and documenting
changes and the necessary resolution is really great, but you only
realize that once other projects don't follow it as religiously.